소스 최신화

+++ cores/domain/src/main/java/com/theloungemembers/core/admin/repository/AdminMenuJpaRepository.java
@@ -5,4 +5,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import com.theloungemembers.core.admin.entity.AdminMenuEntity;
 
-public interface AdminMenuJpaRepository extends JpaRepository<AdminMenuEntity, Integer>, JpaSpecificationExecutor<AdminMenuEntity> {}
+public interface AdminMenuJpaRepository
+        extends JpaRepository<AdminMenuEntity, Long>, JpaSpecificationExecutor<AdminMenuEntity> {
+}
